+++ react/components/Input/index.js
@@ -47,10 +47,25 @@ class Input extends Component {
       helpText,
       dataAttributes,
       prefix,
+      suffix: suffixProp,
       suffixIcon,
       groupBottom,
     } = this.props
     const { active } = this.state
+
+    const suffix = suffixProp || suffixIcon
+
+    if (suffixIcon) {
+      console.warn(
+        'The prop suffixIcon is deprecated and will be removed in the next major. Please use the prop suffix instead.'
+      )
+    }
+
+    if (prefix && suffix) {
+      console.warn(
+        'You should not use both prefix and suffix props in the same input. '
+      )
+    }
 
     const dataAttrs = {}
     for (const key of Object.keys(dataAttributes)) {
@@ -94,25 +109,19 @@ class Input extends Component {
 
     switch (size) {
       case 'large':
-        classes += `f5 pv4 ${prefix ? 'pl8 pr6' : 'ph6'} ${
-          suffixIcon ? 'pr8' : ''
-        }`
+        classes += `f5 pv4 ${prefix ? 'pl8 pr6' : 'ph6'} ${suffix ? 'pr8' : ''}`
         labelClasses += 'f5 '
         prefixAndSuffixClasses += 'ph4 f5'
         // iconSize = 18
         break
       case 'x-large':
-        classes += `f4 pv5 ${prefix ? 'pl8 pr7' : 'ph7'} ${
-          suffixIcon ? 'pr8' : ''
-        }`
+        classes += `f4 pv5 ${prefix ? 'pl8 pr7' : 'ph7'} ${suffix ? 'pr8' : ''}`
         labelClasses += 'f5 '
         prefixAndSuffixClasses += 'ph5 f4 '
         // iconSize = 22
         break
       default:
-        classes += `f6 pv3 ${prefix ? 'pl7 pr5' : 'ph5'} ${
-          suffixIcon ? 'pr7' : ''
-        }`
+        classes += `f6 pv3 ${prefix ? 'pl7 pr5' : 'ph5'} ${suffix ? 'pr7' : ''}`
         labelClasses += 'f6 '
         prefixAndSuffixClasses += 'ph3 fw5 f6 '
         // iconSize = 16
@@ -174,7 +183,7 @@ class Input extends Component {
               WebkitAppearance: 'none',
             }}
           />
-          {suffixIcon && (
+          {suffix && (
             <span
               style={{
                 height: calcPrefixAndSuffixHeight,
@@ -182,7 +191,7 @@ class Input extends Component {
                 right: prefixAndSuffixPosition,
               }}
               className={prefixAndSuffixClasses}>
-              {suffixIcon}
+              {suffix}
             </span>
           )}
         </div>
@@ -231,7 +240,7 @@ InputWithRef.propTypes = {
   /** Label */
   label: PropTypes.string,
   /** Prefix */
-  prefix: PropTypes.string,
+  prefix: PropTypes.node,
   /** Internal prop used for ref forwarding */
   forwardedRef: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
   /** Spec attribute */
@@ -284,7 +293,9 @@ InputWithRef.propTypes = {
   src: PropTypes.string,
   /** Spec attribute */
   step: PropTypes.string,
-  /** SuffixIcon attribute */
+  /** Suffix attribute */
+  suffix: PropTypes.node,
+  /** DEPRECATED: Suffix icon attribute */
   suffixIcon: PropTypes.element,
   /** Spec attribute */
   tabIndex: PropTypes.string,